changed memkind prefix to 1 byte

+++ src/object.c
@@ -429,7 +429,7 @@ void *dupObjectPM(robj* o) {
     if (!o->ptr) return o;
     if (o->encoding == OBJ_ENCODING_INT) return o;
 
-    uint64_t *is_ram = (uint64_t*)((char*)(o->ptr) - MEMKIND_PREFIX_SIZE);
+    uint8_t *is_ram = (uint8_t*)((char*)(o->ptr) - MEMKIND_PREFIX_SIZE);
     if (*is_ram) {
       return o->ptr;
     }
